+++ tests/deprecated_api/test_remove_1-7.py
@@ -39,12 +39,6 @@ from tests.loggers.test_logger import CustomLogger
 from tests.plugins.environments.test_lsf_environment import _make_rankfile
 
 
-def test_v1_7_0_moved_model_summary_and_layer_summary(tmpdir):
-    _soft_unimport_module("pytorch_lightning.core.memory")
-    with pytest.deprecated_call(match="to `pytorch_lightning.utilities.model_summary` since v1.5"):
-        from pytorch_lightning.core.memory import LayerSummary, ModelSummary  # noqa: F401
-
-
 def test_v1_7_0_datamodule_transform_properties(tmpdir):
     dm = MNISTDataModule()
     with pytest.deprecated_call(match=r"DataModule property `val_transforms` was deprecated in v1.5"):